Raul

+++ Ademicon/Ademicon/src/main/resources/db/migration/V8__create_table_propostas.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS propostas(
     data_visita TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     retorno_previsto DATE NOT NULL DEFAULT CURRENT_DATE,
     categoria_proposta VARCHAR(50) NOT NULL,
-    valor_proposta INTEGER,
+    valor_proposta DECIMAL,
     campo_observacao VARCHAR(500),
     id_status INTEGER NOT NULL,
     FOREIGN KEY(id_status) REFERENCES status(id_status)
